adds default admin home at install and update
