+++ main/core/Library/Installation/Updater/Updater120501.php
@@ -11,9 +11,7 @@
 
 namespace Claroline\CoreBundle\Library\Installation\Updater;
 
-use Claroline\AppBundle\API\SerializerProvider;
 use Claroline\AppBundle\Persistence\ObjectManager;
-use Claroline\CoreBundle\Entity\Tab\HomeTab;
 use Claroline\InstallationBundle\Updater\Updater;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -24,8 +22,6 @@ class Updater120501 extends Updater
     private $conn;
     /** @var ObjectManager */
     private $om;
-    /** @var SerializerProvider */
-    private $serializer;
 
     public function __construct(ContainerInterface $container, $logger = null)
     {
@@ -33,14 +29,12 @@ class Updater120501 extends Updater
         $this->container = $container;
         $this->conn = $container->get('doctrine.dbal.default_connection');
         $this->om = $container->get('claroline.persistence.object_manager');
-        $this->serializer = $container->get('claroline.api.serializer');
     }
 
     public function postUpdate()
     {
         $this->updateWorkspaceRedirection();
         $this->removeTool('dashboard');
-        $this->addDefaultAdminWidget();
     }
 
     private function updateWorkspaceRedirection()
@@ -78,49 +72,5 @@ class Updater120501 extends Updater
         $this->log($sql);
         $stmt = $this->conn->prepare($sql);
         $stmt->execute();
-    }
-
-    private function addDefaultAdminWidget()
-    {
-        $tabs = $this->om->getRepository(HomeTab::class)->findBy(['type' => HomeTab::TYPE_ADMIN]);
-        if (!empty($tabs)) {
-            $serializedTab = $this->serializer->serialize($tabs[0]);
-
-            if (!isset($serializedTab['widgets'])) {
-                $serializedTab['widgets'] = [];
-            }
-
-            // adds Admin Tools widget (it's easier to use JSON, to be sure all Entities are correctly created)
-            $serializedTab['widgets'][] = [
-                'visible' => true,
-                'display' => [
-                    'layout' => [1],
-                    'color' => '#333333',
-                    'backgroundType' => 'color',
-                    'background' => '#ffffff',
-                ],
-                'parameters' => [],
-                'contents' => [[
-                    'type' => 'list',
-                    'source' => 'admin_tools',
-                    'parameters' => [
-                        'showResourceHeader' => false,
-                        'display' => 'tiles-sm',
-                        'enableDisplays' => false,
-                        'availableDisplays' => [],
-                        'card' => [
-                            'display' => ['icon', 'flags', 'subtitle'],
-                        ],
-                        'paginated' => false,
-                        'count' => false,
-                    ],
-                ]],
-            ];
-
-            $tab = $this->serializer->deserialize($serializedTab, $tabs[0]);
-
-            $this->om->persist($tab);
-            $this->om->flush();
-        }
     }
 }
